requirements.txt
flask>=3.0.2
flask-cors>=4.0.0
openai>=1.12.0
anthropic>=0.18.1
beautifulsoup4>=4.12.3
selenium>=4.18.1
python-dotenv>=1.0.1
requests>=2.31.0
typing-extensions>=4.9.0
Pillow>=10.0.0
Flask-SQLAlchemy>=3.0.2
PyJWT>=2.6.0
cryptography>=41.0.1
pyOpenSSL>=23.2.0
Werkzeug>=3.0.0
Flask-Limiter>=3.3.1
Flask-Talisman>=1.0.0
Flask-SeaSurf>=1.1.1
Flask-WTF>=1.1.1
bleach>=6.0.0 